test: use nock for credential revocation webhook
